feat(PPDSC-2064): feat-ppdse/2064 created menu in site
